+++ presentation/src/test/java/com/agileburo/anytype/presentation/auth/SelectAccountViewModelTest.kt
@@ -2,21 +2,14 @@ package com.agileburo.anytype.presentation.auth
 
 import MockDataFactory
 import androidx.arch.core.executor.testing.InstantTaskExecutorRule
-import androidx.lifecycle.viewModelScope
 import com.agileburo.anytype.domain.auth.interactor.ObserveAccounts
 import com.agileburo.anytype.domain.auth.interactor.StartLoadingAccounts
 import com.agileburo.anytype.domain.auth.model.Account
-import com.agileburo.anytype.domain.base.Either
 import com.agileburo.anytype.presentation.auth.account.SelectAccountViewModel
-import com.agileburo.anytype.presentation.auth.model.SelectAccountView
 import com.agileburo.anytype.presentation.util.CoroutinesTestRule
 import com.jraska.livedata.test
-import com.nhaarman.mockitokotlin2.any
-import com.nhaarman.mockitokotlin2.doAnswer
 import com.nhaarman.mockitokotlin2.doReturn
 import com.nhaarman.mockitokotlin2.stub
-import kotlinx.coroutines.cancel
-import kotlinx.coroutines.delay
 import kotlinx.coroutines.flow.asFlow
 import kotlinx.coroutines.flow.flow
 import kotlinx.coroutines.test.runBlockingTest
@@ -25,7 +18,6 @@ import org.junit.Rule
 import org.junit.Test
 import org.mockito.Mock
 import org.mockito.MockitoAnnotations
-import org.mockito.MockitoDebugger
 
 class SelectAccountViewModelTest {
 
@@ -56,7 +48,7 @@ class SelectAccountViewModelTest {
     }
 
     @Test
-    fun `should emit one account without image`() = runBlockingTest {
+    fun `should not emit one account without image`() = runBlockingTest {
 
         val account = Account(
             id = MockDataFactory.randomUuid(),
@@ -74,21 +66,11 @@ class SelectAccountViewModelTest {
 
         vm = buildViewModel()
 
-        vm.state.test()
-            .assertValue(
-                listOf(
-                    SelectAccountView.AccountView(
-                        id = account.id,
-                        name = account.name,
-                        image = account.avatar
-                    )
-                )
-            )
-            .assertHistorySize(1)
+        vm.state.test().assertNoValue()
     }
 
     @Test
-    fun `should emit list with two accounts without images`() = runBlockingTest {
+    fun `should not emit list with two accounts without images`() = runBlockingTest {
 
         val firstAccount = Account(
             id = MockDataFactory.randomUuid(),
@@ -109,30 +91,12 @@ class SelectAccountViewModelTest {
             emit(secondAccount)
         }
 
-        val blob = ByteArray(0)
-
-        val response = Either.Right(blob)
-
         observeAccounts.stub {
             onBlocking { build() } doReturn accounts
         }
 
         vm = buildViewModel()
 
-        vm.state.test()
-            .assertValue(
-                listOf(
-                    SelectAccountView.AccountView(
-                        id = firstAccount.id,
-                        name = firstAccount.name
-                    ),
-                    SelectAccountView.AccountView(
-                        id = secondAccount.id,
-                        name = secondAccount.name
-                    )
-                )
-            )
-            .assertHistorySize(1)
-
+        vm.state.test().assertNoValue()
     }
 }